+++ AdhereR/NEWS.md
@@ -1,4 +1,10 @@
-# AdhereR 0.7.1
+# AdhereR 0.8.1
+  
+## Bug fixes
+
+  - fix `CMA_polypharmacy()` crash due to not properly handling the new internal `suppress.special.argument.checks` argument (now there's also a unit test for `CMA_polypharmacy` to avoid such bugs in the future).
+
+# AdhereR 0.8.0
   
 ## New features